added gpt training shell script
